+++ src/SonnetDB/Json/ServerJsonContext.cs
@@ -2,7 +2,9 @@
 using System.Text.Json.Serialization;
 using SonnetDB.Contracts;
 using SonnetDB.Copilot;
+using SonnetDB.Documents;
 using SonnetDB.Graphs;
+using SonnetDB.FullText;
 using SonnetDB.Mcp;
 using SonnetDB.Modbus;
 using SonnetDB.Model;
@@ -260,6 +262,17 @@ namespace SonnetDB.Json;
 [JsonSerializable(typeof(FullTextAnalyzeResponse))]
 [JsonSerializable(typeof(FullTextTokenInfo))]
 [JsonSerializable(typeof(List<FullTextTokenInfo>))]
+[JsonSerializable(typeof(FullTextSettingsRequest))]
+[JsonSerializable(typeof(DocumentFullTextIndexSettings))]
+[JsonSerializable(typeof(DocumentFullTextTypoPolicy))]
+[JsonSerializable(typeof(FullTextAnalyzerDiffRequest))]
+[JsonSerializable(typeof(FullTextAnalyzerDiffResponse))]
+[JsonSerializable(typeof(FullTextRelevanceExplainRequest))]
+[JsonSerializable(typeof(DocumentFullTextRelevanceExplanation))]
+[JsonSerializable(typeof(DocumentFullTextTermContribution))]
+[JsonSerializable(typeof(List<DocumentFullTextTermContribution>))]
+[JsonSerializable(typeof(FullTextRebuildRequest))]
+[JsonSerializable(typeof(FullTextRebuildStatusResponse))]
 [JsonSerializable(typeof(MqTopicListResponse))]
 [JsonSerializable(typeof(MqTopicInfo))]
 [JsonSerializable(typeof(List<MqTopicInfo>))]